Fix typo CRITCAL > CRITICAL
I noticed a probelm in CheckNSCP.cpp and CheckHelpers that
sometimes CRITCAL was being returned as a status.

Not sure if it was intentional or not, but I replaced all
occurances.

It looks like some of the source files have old DOS file endings
(^M) instead of unix-friendly linebreaks. My search replace stripped
some of them and I hope that does not cause any damage.

Please check it out and let me know if this was the right thing
to do.